Fix: Compile Error

+++ includes/object.h
@@ -6,7 +6,7 @@
 /*   By: jukim2 <jukim2@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/30 16:45:23 by kjs               #+#    #+#             */
-/*   Updated: 2023/10/03 21:25:59 by jukim2           ###   ########.fr       */
+/*   Updated: 2023/10/03 21:35:08 by jukim2           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_ambient_light
 {
-	float		light_ratio;
+	double		light_ratio;
 	t_vector3	rgb;
 } t_ambient_light;
 
@@ -25,13 +25,13 @@ typedef struct s_camera
 {
 	t_vector3	coord;
 	t_vector3	forward_vector;
-	float		fov;
-	// Not sure fov is float or int
+	double		fov;
+	// Not sure fov is double or int
 } t_camera;
 
 typedef struct s_light
 {
-	float		light_ratio;
+	double		light_ratio;
 	t_vector3	coord;
 	t_vector3	rgb;
 } t_light;
@@ -40,8 +40,8 @@ typedef struct s_shape
 {
 	struct s_shape	*next;
 	int				type;
-	float			diameter;
-	float			height;
+	double			diameter;
+	double			height;
 	t_vector3		coord;
 	t_vector3		normal_vector;
 	t_vector3		rgb;
